requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
Django==3.2.1
django-graphql-jwt 0.3.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
pytest==6.2.4
pytest-django==4.4.0
pytz==2021.1
Rx==1.6.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
